+++ tinper-bee/bee-colorpicker/demo/ColorpickerDemo.js
@@ -1,0 +1,5 @@
+import ColorPicker from '../src/index';
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+class Demo extends Component {render(){return( <ColorPicker/> )}}
+export default Demo;