+++ tinper-bee/bee-datepicker/CHANGELOG.md
@@ -1,3 +1,28 @@
+<a name="2.0.16"></a>
+## [2.0.16](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.15...v2.0.16) (2019-03-26)
+
+
+### Bug Fixes
+
+* **bee-datepicker:** 去掉多余console ([26cfb0d](https://github.com/tinper-bee/bee-datepicker/commit/26cfb0d))
+
+
+
+<a name="2.0.15"></a>
+## [2.0.15](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.15-beta...v2.0.15) (2019-03-26)
+
+
+### Bug Fixes
+
+* **bee-datepicker:** openchange ([d42d17b](https://github.com/tinper-bee/bee-datepicker/commit/d42d17b))
+
+
+
+<a name="2.0.15-beta"></a>
+## [2.0.15-beta](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.14...v2.0.15-beta) (2019-03-26)
+
+
+
 <a name="2.0.14"></a>
 ## [2.0.14](https://github.com/tinper-bee/bee-datepicker/compare/v2.0.13...v2.0.14) (2019-03-26)